Try to fix the trace output when the lisp_print trace option is set
The last line of the trace output could run into the output of the
result of evaluating the expression.  For results displayed in 2D,
this would screw up the alignment of the output.

Also if multiple functions were actively being traced, and the
lisp_print trace option was set for some (but not all) of them, the
regular trace output for one function could be on the same line as
the Lisp object trace output for the previous function.

Now print a newline (instead of a space) after a Lisp object trace
line has been printed (this seems to be consistent with the regular
trace output).

Also do not print a newline at all before the Lisp object trace line
is printed since this causes a superfluous newline at the beginning
of all of the trace output that is not present when using the regular
trace output.  This would also cause superfluous newlines to appear
when mixing regular and Lisp object trace outputs when display2d was
set to false.