Fix evaluation of li[3](z) for z near the unit circle.
The previous algorithm would get in a loop evaluating points on or
near the unit circle. The algorithm would change the arg to another
point on or near the circle which would require another arg change
going back to the original.

Break the cycle by implementing a new algorithm when the arg is near
the unit circle.  The algorithm is from cephes which gives an
expansion of li[s] in terms of log.

Tests added for this; the test values are from functions.wolfram.com.