Split up implementation of FFT in a separate package
Move all of the functions that are exposed to Maxima to the (new) file
fft-interface.lisp.  The core internal algorithms remain in
fft-core.lisp but are now in the MAXIMA-FFT package instead of MAXIMA.

Add fft-package.lisp to define the MAXIMA-FFT package, exporting the
intended interface from the MAXIMA-FFT package to callers.

All fft tests pass.