Eliminate a superfluous PROG1
In v5.0, this PROG1 was used to call MUNLOCAL after the LET form
(with a comment saying this MUNLOCAL should be unwind-protected
instead).  At some point prior to the initial repo commit, some
things were redone, UNWIND-PROTECT was used to call MUNLOCAL
instead and the MUNLOCAL call in the PROG1 was replaced with NIL.

No functional changes.