Cut out the broken and unused macro DEF$FUN in share/affine/polyd.lisp
This macro was supposed to be for defining functions and automatically
declaring any variables in the lambda list beginning with a dollar sign
to be special.  However, this was broken in CL because it would use
CHAR on a symbol (not the symbol's name), which would cause an error.

In commit 7ccc7f the last use of this macro was removed (I'm guessing
that the intention was to not use it on CL implementations because the
DEF$FUN form was subject to a read-time test #-cl, but that's not a
standard feature... this code was from 1996 or before).  The only
other occurrences of this macro have been examples in comments.

Also cut out the related and otherwise unused special variable
*SHOW-SPECIALS*.