No longer compare characters with EQ
The standard doesn't guarantee that characters with the same value are
EQ, so compare with EQL instead.  In some cases CHAR= could be used
(because CHAR= and EQL compare characters in the same way), but in
some other cases CHAR= wouldn't work (like when GETCHARN is involved,
since it can return NIL instead of a character).

This is related to my series of commits regarding the use of EQ for
comparing numbers.