Remove an outdated (incorrect in CL) comment in src/limit.lisp
This comment is present in the Maxima v5.0 tarball and the last
modification date of src/limit.lisp in there is 1987 (although I
have no idea when this comment was actually added).  I'm guessing
that this comment is about an older lisp dialect because there is
nothing special in CL about using NIL as a catch tag.