No longer compare numbers with EQ [part 4]
The standard doesn't guarantee that numbers with the same value are
EQ, so compare with EQL instead.

This is just taking care of some leftovers that I've found.  The
change in src/commac.lisp is because in at least some cases EOF-OPTION
is *PARSE-STREAM-EOF*, which is -1.  It's seems that the intention of
the code in src/pois3.lisp is to compare just the symbols in that list
(because of the CADR), but I'm changing the EQ test anyway.

See also commits dac725, 34f1f8 and 050553 (parts 1, 2 and 3).