Mark tests that require the docs as known failures in Lisp-only builds
Two tests in rtest5 depend on the docs being present, so they would
fail if the docs were missing.  The lisp-only build procedure doesn't
build the docs, so mark these tests as known failures in that case
(the tests will be marked as known failures and pass anyway if the
docs are actually available).