Fix an egrep pattern which broke the finding of the share directories
My understanding is that the goal is to match lines that contain ".t2p"
for default_sharedirs.

The egrep pattern introduced in commit 012756 is "*.t2p".  The egrep on
OpenBSD (and most likely others) throws an error because of the
placement of the asterisk, which makes default_sharedirs empty, which
makes it so none of the share directories can be found.

Now use "\.t2p" instead.  This one matches a literal period.

Tested on OpenBSD and Linux using some artificial .t2p directories
in share.