Fix #3186 and other get_plot_option issues.
First, fix #3186 so that the last entry of *plot-options* is actually
returned.

Second, make get_plot_option return a copy of *plot-options*;
previously, the result would share structure (like the values for
color) with *plot-options*.  Then, modifications to the result would
modify *plot-options*.  This seems to be against the intent of
*plot-options*.

Third, replace the O(n^2) algorithm for copying *plot-options* with
a simple O(n) algorithm by scanning *plot-options* in order, stepping
by cddr.  This isn't a big deal since the list is short, this change
makes the algorithm a bit clearer.

Finally, add a few tests for this.