Bug fix: now things like fourier_elim(2^x < 2^y, [x, y]) yield [x < y]
There were some problems in the COND clause in M> which is supposed to
handle "Do a^x > a^y, where a > 1 --> x > y":

* it was potentially comparing numbers (the bases of the exponentials)
  with EQ, which is not guaranteed to work.  Now compare these with
  EQL instead (this could be relaxed further, if desired).

* it was comparing strings (the result from $COMPARE and ">") with EQ,
  which would not work in general (these strings would not "normally"
  be EQ, although it's possible for them to be if the compiler decides
  that they can be "collapsed").  Now compare these with EQUAL instead.

* it was comparing the bases of the exponentials a second time (this
  time with M>) rather than comparing the exponents.

In "most" cases (or, all of the ways that I tried) the strings would
not be EQ, which would mean that the example fourier_elim call above
would yield [2^y - 2^x > 0].  Otherwise, if the number and string
comparisons worked as intended, then the fourier_elim call would
erroneously yield emptyset (since the bases of the exponentials would
be EQ, a M> comparison of them would yield NIL).

rtest_fourier_elim runs fine, with two new tests.