Now maxima can directly load .wxm files from wxMaxima
.wxm files contain all comments and sectioning commands that are needed
in order to reproduce wxMaxima documents. And they contain all of the maxima
code the document contains.

Since in a .wxm fole everything except the code is wrapped between
/* and */ maxima can interpret the result of saving a .wxm file from
wxMaxima without making any big changes to the code. All that needed
to be done in order to make this work was to tell maxima that it actually
is able to handle the new file type.