SF [bug:#1266][bug:#3208] Don't use $radcan when simplifying in algsys.
Inappropriate use of $radcan when simplifying expressions caused a
number of failures in algsys.  This commit follows on from recent
work that removed one call to radcan and introduced use of sqrtdenest.
This addresses a number of long-standing solve/algsys failures.

The recent changes to ebaksubst simpified some expressions to zero
and allowed the solution to continue until it crashed in a call to
radcan from pfreeofmainvarsp.

* src/algsys.lisp:
  New function simplify-after-subst that uses sqrtdenest, $rectform
  and $ratsimp.  This replaces uses of radcan to simplify expressions.
  Call simplify-after-subst from ebaksubst1, pfreeofmainvarsp and
  pfreeofmainvarsp.

* tests/rtest8.mac: Change form of one solution
* tests/rtest_algsys.mac: Update test results.  Add new tests.

* share/contrib/diffequations/tests
   New solutions from odelin due to algsys improvements.
   rtestode_kamke_1_2.mac:  Kamke 1.105
   rtestode_kamke_2_3.mac:  Kamke 2.275 and 2.292
   rtestode_murphy_1_2.mac: Murphy 1.182