Complex numerical eval returns args
complex-numerical-eval-p (for floats and bigfloats) basically computes
the rectform of each arg and determines if the arguments are numerical
so that the function should be evaluated numerically.

The computation of the rectform is relatively expensive, so it's
useful if this function returns the reduced args, so make the function
return the args if the conditions are satisfied (or $numer is true).

Then the caller can extract out the reduced values for the computation
instead of computing the rectform itself again.