A solve() enhancement by Richard Fateman that should fix a zillion of bug tickets.
One test case that is resolved by this commit is:

eqs:[(- a*n2) - a*n1 - 2*a, sqrt(a)*sqrt(- c)*(n2 - n1), b*n4 + n3
+ b*n2 + b*n1 + b + 1];
	vars:[n1,n2,n3,n4];
	solve(eqs,vars);

The original mail is:
recently I suggested that solve(eq3,vars)  was a good test case.

I changed the definition of backward in mat.lisp
so that instead of computing
    (pquotient  mess1 mess2)

it first checked to see if mess1 or mess2 was zero in which case
returning 0 otherwise computing
the quotient.

If this is all that is required in other cases where pquotient is
called, that would be simple enough.
another possibility would be to do that testing and also use
(car(ratreduce mess1 mess2)) instead of (pquotient mess1 mess2)

Anyway, this fixes the solve example.  Try reading it in to your maxima,
starting at the line below.  Someone -- please patch the source.

The whole discussion thread can be viewed at https://sourceforge.net/p/maxima/mailman/message/35594033/