Add a fallback for %MAXIMA_TEMPDIR% to maxima.bat
There were reports, that the directory could not be found.
Also be sure to enable cmd extensions. They are enabled by
default, but may be disabled in the registry.

Changed the test if %MAXIMA_USERDIR% and %MAXIMA_TEMPDIR% exists to
if not defined MAXIMA_USERDIR
That seems to work better.
if "%MAXIMA_USERDIR%" == ""
does not work if the variable contains quotes (and maybe other issues too?)

Maybe the other tests, if variables exists should also be changed - but should
I change it, if it is not (reported) broken?