Documentation: Removed @urefbreakstyle which prevented a build on the mac I tried it on.
Background info from Robert Dodier:

A bit of searching the web shows that @urefbreakstyle was introduced in
Texinfo 5.0 which was in February 2013. I see a few options (and no
doubt there are others).

 * require Texinfo 5.0 or later
 * try to detect the Texinfo version and use the command only if version
   is recent enough
 * cut out the use of the problematic command

My recommendation at this point is to put the URLs which are mentioned
into footnotes instead of the body of the text, and then cut out
@urefbreakstyle. I'd prefer to have URLs in footnotes anyway; URLs
clutter the text.