mimetypes: A mimetype for xmaxima's .out files.
wxMaxima now supports extracting the output of xmaxima's .out files - at
least in the case that the .out file format allows this (disp and printf
commands can look like just another line of input). In order to allow
wxMaxima to open the .out files they need to be assigned a mimetype to.
In theory this could have been done on wxMaxima's side. But as it is xmaxima
that creates these files it felt more logical to make maxima do this,
instead.

In order to tell .out files created by xmaxima from .out files created by
compilers or other programs we only claim the file to be generated by
xmaxima if the file begins with "(%i1) ".