xmaxima: standard help message behavior
On encountering the command-line arguments "-h" or "-help" xmaxima
created a transient window containing a help message - which was nice,
but nonstandard => Changed the behaviour to outputting a help message
on stdout on encountering the arguments "-h", "--help" or "-help".

Rationale: This will enable the possibility to make bash (and possibly
others) automatically parse the help message.