Documentation: A cross-reference from ev() to subst() and at().
We keep telling people not to use ev(). But we didn't link from ev()
to alternatives that are appropriate in common cases ev() is
mis-used for.

In the pdf version of the manual @mxrefcomma{} produces one comma
too many for me. I assume this to be a bug in my texinfo version.