Move the computation of the Jacobian to the init function.
Instead of computing the Jacobian in the step function every time we
do a time step, compute it in the init function, save it in the state,
and use that in the step function.

This also means that the vars variable for the init function must
contain the independent variable too.