Crosscompiling: update Imagemagick
They update very often (and remove old releases), so I have to update it
here often too.
Changed the hash from MD5 to SHA256. No, *not* for 'security' reasons.
The hashes for the downloads are for efficiency, so that during development
a package is not downloaded more often then necessary (what happened very
often at the beginning, when I developed the crosscompiling procedure)
But a SHA256 hash is provided by them:
https://www.imagemagick.org/download/binaries/digest.rdf
(at on the same server as the downloads, so an 'attacker' could change a
download and the hash file, so no security involved), and I can use
their precompiled hashes.