Perform more checks for malformed lambda expressions during translation
Previously the following things in lambda expressions were checked
for in the interpreter but not the translator:

  * invalid parameter list -- lambda(atom) would cause a lisp error
  * an invalid "rest" parameter with multiple elements
  * empty body

Now these are checked for during translation as well.  See also
commits f73f82 and 08d189, which add other checks for lambda
expressions to the translator.

More tests have been added to rtest_translator.