Add a comment and a warning message to the DEF%TR for MDEFMACRO
The comment and warning message mention that there is a side-effect
when translating embedded macro definitions: the macro is defined
during translation without being cleaned up.  It's a bug that this
is not cleaned up, but in many cases (as Stavros Macrakis noted)
simply removing this causes the macro to not be defined when it's
time to do the macro expansion.