Fix/improve how create_list binds its iteration variables
This addresses most of the concerns in bug #3220.  Now

* setcheck tracing works for the iteration variables
* constants cannot be used as iteration variables
* using fpprec as an iteration variable works correctly

The only remaining concern from the bug report is that iteration
variables are still required to be symbols (which is documented
behavior, for whatever that's worth).

The test suite runs fine, with new tests.  No changes in the
share test suite.