Avoid the endless loop of ECL if wxMaxima crashes
wxMaxima is connected to maxima via a network socket - which is the
standard way to connect to maxima. If wxMaxima crashes inexpectedly this
causes maxima to issue an error message - which normally caused an error
(as the error message was sent over the now-extinct network connection)
causing an endles loop that ate up 100% of one CPU => Added code that
causes maxima to quit if outputting an error message causes another error.