Windows: Add a workaround for Xmaxima / Lisp
That combination did not work, because %dp0 was interpreted as a
short (8.3) filename and Xmaxima did not work. If Maxima is installed
in the default path (which *is* recommended) use the path for which Maxima
was configured. Xmaxima works then. Xmaxima still does not work if it is
installed somewhere else.
Command line Maxima and wxMaxima do not have these troubles, they work
in both cases.