A change in the name of the functions
The functions that work with permutations, given as lists of consecutive
integers from 1 to n, will now use the prefix "perm", to distinguish
them from those that permute any arbitrary list or set, which use the
prefix "permutation". The name of permutation_index was changed to
perm_length and apply_permutation became permute.