"make check" now runs the share tests.
We still don't run the share and the regular testbench in the same maxima
session as that causes additional test failures. But this way we at least
get feedback from the nightly build if anything changes in maxima that
causes something to fail without having to bisect it:
Having caused about 200 test failure the share testbench seems to have
been subject to a lot of bit rot in the last few years.

Additionally fixed the bug that ccl64 if "make check" ran multiple jobs
in parallel would sometimes fail because it used the wrong log file.