"make check": (gcl only): Ensure that gcl uses sane memory management settings.
The versions of gcl that can be found in Debian and Ubuntu currently
need loads of Gigabytes on running "make check". There is a known
workaround. This commit adds it to "make check", but not to the
stock maxima as I don't know if it has any side effects the users
would notice.