A simple but powerful tolerance calculation package for maxima.
Maxima is used by many engineers. It is mathematics cast in
silicon and if an engineer uses mathematics then either in order
to get to understand a mechanism or in order to do worst-case
calculations => maxima needs a worst-case calculation package and
it needs to be part of maxima's core.

Also I want to be able to distribute it at work in the next few
weeks which means it needs to be part of the upcoming maxima
release.

Tried to make the interface simple enough that it won't occupy
more space in maxima's manual than absolutely needed; All
user-accessible functions are prefixed with "wc_" which makes
them easier to find in wxMaxima's autocompletion: If it is
worst-case typing
  wc <Shift+Tab>
will find it.

The package runs reasonably fast. What it cannot provide is a
complete interval arithmetics (which is the reason the name
is shortened in respect to "worstcase"): It actually would be
nice to have such a functionality. But I guess interval
arithmetics would be subject to the halting problem anyway and
hence there still will be cases in which an approach that just
tries random values will still be handy.

It is astonishing with how little a number of tries
with random tolerance values a reasonable gaussian distribution
and reasonable min and max values can be found even in complex
cases that use many parameters with tolerances. In theory one
could construct a case that hides a maximum that is narrow
enough so no guess will ever hit it. But engineering doesn't
tend to go near such a place.

The most severe limitation of this package therefore seems to
be that min() and max() sometimes run out of time and/or memory
in places where a human being would still be able to guess the
outcome.

As I tend to get questions if guessing the maximum using the
gradients of an expression would be more powerful: In
electrotechnics this approach seems to be more prone to find
a local optimum than the one used in this package.
I hope this to be true in most fields of engineering.