Fix charfun when prederror is true
charfun is documented to return a nounform when the predicate
evaluates to something other than true or false; however, if
prederror was true then an error could occur instead.

Now bind prederror to false before calling MEVALP.

The test suite runs fine (with a new test).  No changes in the
share test suite.