Add a warning() command:
I get frequently asked if a worksheet can output warnings as well as
errors if a tolerance calculation doesn't fail but yield an unexpected
result. And if this warning can be output in a way that automatically
causes it to be formatted by wxMaxima as a warning.

This command matches both needs.