In ALIST-BIND, declare the variable holding the alist to be ignorable
If the variable list given to ALIST-BIND is empty, then some lisp
implementations issue a warning about a gensym being defined but
unused.  This gensym is used by ALIST-BIND to hold the alist, but
it is otherwise unused if the variable list is empty.

ALIST-BIND is used in SCHATCHEN-COND and the documented way to get
an unconditional match without binding any variables here is to use

  (schatchen-cond w
    ...
    (t nil ...))

where nil is the variable list.  In this case ALIST-BIND is used on
this empty variable list.

SP1TRIGEX uses SCHATCHEN-COND in this way, which resulted in the
non-obvious warning message about the unused gensym.  The change
to use SCHATCHEN-COND in SP1TRIGEX was in commit 6db8907f and the
introduction of ALIST-BIND was in commit c2dcbeed.

And anyway, if you explicitly ask ALIST-BIND to not create any
bindings then I don't think you should get a mysterious warning
about an unused gensym.