Fix bug #3368: integrate('limit(...),...) internal error
No cons check was being performed before taking the CAAR of an
expression in INTEGRATOR, so we'd get a lisp error in some cases.

The original example from Stavros Macrakis:

  integrate('limit(x,x,0),x)      =>  lisp error

Here are some others that I've found:

  integrate('at(x,x=1),x)         =>  lisp error
  integrate('sum(x,x,1,1),x)      =>  lisp error
  integrate('product(x,x,1,1),x)  =>  lisp error

The proposed fix is also from Stavros (I just used MTIMESP for
consistency with another check in INTEGRATOR).

The test suite runs fine, with new tests.