Fix a lisp error when display2d:true and there are nested DISPLA calls
LINEARRAY is an array used in the display code.  Every element should
be a list and it should be initialized to all NILs (this is assumed in
the display code).  LINEARRAY is initialized to all NILs when it is
defined, and it is filled with NILs when leaving DISPLA; however, if
DISPLA got called while still in another DISPLA call then LINEARRAY
was getting bound to a new array with uninitialized elements.

CLHS says that (in this case) consequences are undefined when reading
these uninitialized elements.  It looks like ECL 16.1.3, ABCL 1.5.0 and
Clisp 2.49 all make these elements NIL, and all is well in the display
code.  But it looks like at least SBCL 1.4.5 makes these elements 0,
which later causes a lisp error because the display code assumes that
it is safe to call CAR and CDR on these elements.

Now the array created in DISPLA is initialized to all NILs.

This use of an array with uninitialized elements goes at least as far
back as Maxima v5.0.  It looks like LINEARRAY itself was defined but
uninitialized prior to commit 00c8f1e3.

I stumbled onto this by tracing a function that was called in a
DISPLA-DEF dispatch function.

The test suite runs fine.